refactoring round
