+++ lib/isaac.go
@@ -4,171 +4,9 @@ import (
 	"errors"
 
 	"boscoin.io/sebak/lib/common"
-	"boscoin.io/sebak/lib/error"
 	"boscoin.io/sebak/lib/node"
+	"boscoin.io/sebak/lib/round"
 )
-
-type RoundVoteResult map[ /* Node.Address() */ string]sebakcommon.VotingHole
-
-type RoundVote struct {
-	SIGN   RoundVoteResult
-	ACCEPT RoundVoteResult
-}
-
-func NewRoundVote(ballot Ballot) (rv *RoundVote) {
-	rv = &RoundVote{
-		SIGN:   RoundVoteResult{},
-		ACCEPT: RoundVoteResult{},
-	}
-
-	rv.Vote(ballot)
-
-	return rv
-}
-
-func (rv *RoundVote) IsVoted(ballot Ballot) bool {
-	result := rv.GetResult(ballot.State())
-
-	_, found := result[ballot.Source()]
-	return found
-}
-
-func (rv *RoundVote) IsVotedByNode(state sebakcommon.BallotState, node string) bool {
-	result := rv.GetResult(state)
-
-	_, found := result[node]
-	return found
-}
-
-func (rv *RoundVote) Vote(ballot Ballot) (isNew bool, err error) {
-	if ballot.IsFromProposer() {
-		return
-	}
-
-	result := rv.GetResult(ballot.State())
-	_, isNew = result[ballot.Source()]
-	result[ballot.Source()] = ballot.Vote()
-
-	return
-}
-
-func (rv *RoundVote) GetResult(state sebakcommon.BallotState) (result RoundVoteResult) {
-	if !state.IsValidForVote() {
-		return
-	}
-
-	switch state {
-	case sebakcommon.BallotStateSIGN:
-		result = rv.SIGN
-	case sebakcommon.BallotStateACCEPT:
-		result = rv.ACCEPT
-	}
-
-	return result
-}
-
-func (rv *RoundVote) CanGetVotingResult(policy sebakcommon.VotingThresholdPolicy, state sebakcommon.BallotState) (RoundVoteResult, sebakcommon.VotingHole, bool) {
-	threshold := policy.Threshold(state)
-	if threshold < 1 {
-		return RoundVoteResult{}, sebakcommon.VotingNOTYET, false
-	}
-
-	result := rv.GetResult(state)
-	if len(result) < int(threshold) {
-		return result, sebakcommon.VotingNOTYET, false
-	}
-
-	var yes, no int
-	for _, votingHole := range result {
-		switch votingHole {
-		case sebakcommon.VotingYES:
-			yes++
-		case sebakcommon.VotingNO:
-			no++
-		}
-	}
-
-	log.Debug(
-		"check threshold in isaac",
-		"threshold", threshold,
-		"yes", yes,
-		"no", no,
-		"policy", policy,
-		"state", state,
-	)
-
-	if yes >= threshold {
-		return result, sebakcommon.VotingYES, true
-	} else if no >= threshold {
-		return result, sebakcommon.VotingNO, true
-	}
-
-	// check draw!
-	total := policy.Validators()
-	voted := yes + no
-	if total-voted < threshold-yes && total-voted < threshold-no { // draw
-		return result, sebakcommon.VotingNO, true
-	}
-
-	return result, sebakcommon.VotingNOTYET, false
-}
-
-type RunningRound struct {
-	sebakcommon.SafeLock
-
-	Round        Round
-	Proposer     string                              // LocalNode's `Proposer`
-	Transactions map[ /* Proposer */ string][]string /* Transaction.Hash */
-	Voted        map[ /* Proposer */ string]*RoundVote
-}
-
-func NewRunningRound(proposer string, ballot Ballot) (*RunningRound, error) {
-	transactions := map[string][]string{
-		ballot.Proposer(): ballot.Transactions(),
-	}
-
-	roundVote := NewRoundVote(ballot)
-	voted := map[string]*RoundVote{
-		ballot.Proposer(): roundVote,
-	}
-
-	return &RunningRound{
-		Round:        ballot.Round(),
-		Proposer:     proposer,
-		Transactions: transactions,
-		Voted:        voted,
-	}, nil
-}
-
-func (rr *RunningRound) RoundVote(proposer string) (rv *RoundVote, err error) {
-	var found bool
-	rv, found = rr.Voted[proposer]
-	if !found {
-		err = sebakerror.ErrorRoundVoteNotFound
-		return
-	}
-	return
-}
-
-func (rr *RunningRound) IsVoted(ballot Ballot) bool {
-	roundVote, err := rr.RoundVote(ballot.Proposer())
-	if err != nil {
-		return false
-	}
-
-	return roundVote.IsVoted(ballot)
-}
-
-func (rr *RunningRound) Vote(ballot Ballot) {
-	rr.Lock()
-	defer rr.Unlock()
-
-	if _, found := rr.Voted[ballot.Proposer()]; !found {
-		rr.Voted[ballot.Proposer()] = NewRoundVote(ballot)
-	} else {
-		rr.Voted[ballot.Proposer()].Vote(ballot)
-	}
-}
 
 type TransactionPool struct {
 	sebakcommon.SafeLock
@@ -286,13 +124,13 @@ type ISAAC struct {
 	TransactionPool       *TransactionPool
 	RunningRounds         map[ /* Round.Hash() */ string]*RunningRound
 	LatestConfirmedBlock  Block
-	LatestRound           Round
+	LatestRound           round.Round
 }
 
 func NewISAAC(networkID []byte, node *sebaknode.LocalNode, votingThresholdPolicy sebakcommon.VotingThresholdPolicy) (is *ISAAC, err error) {
 	is = &ISAAC{
-		NetworkID:             networkID,
-		Node:                  node,
+		NetworkID: networkID,
+		Node:      node,
 		VotingThresholdPolicy: votingThresholdPolicy,
 		TransactionPool:       NewTransactionPool(),
 		RunningRounds:         map[string]*RunningRound{},
@@ -301,9 +139,11 @@ func NewISAAC(networkID []byte, node *sebaknode.LocalNode, votingThresholdPolicy
 	return
 }
 
-func (is *ISAAC) CloseConsensus(proposer string, round Round, vh sebakcommon.VotingHole) (err error) {
+func (is *ISAAC) CloseConsensus(proposer string, round round.Round, vh sebakcommon.VotingHole) (err error) {
 	is.Lock()
 	defer is.Unlock()
+
+	is.SetLatestRound(round)
 
 	if vh == sebakcommon.VotingNOTYET {
 		err = errors.New("invalid VotingHole, `VotingNOTYET`")
@@ -342,11 +182,11 @@ func (is *ISAAC) SetLatestConsensusedBlock(block Block) {
 	is.LatestConfirmedBlock = block
 }
 
-func (is *ISAAC) SetLatestRound(round Round) {
+func (is *ISAAC) SetLatestRound(round round.Round) {
 	is.LatestRound = round
 }
 
-func (is *ISAAC) IsAvailableRound(round Round) bool {
+func (is *ISAAC) IsAvailableRound(round round.Round) bool {
 	// check current round is from InitRound
 	if is.LatestRound.BlockHash == "" {
 		return true
