+++ lib/round_vote.go
@@ -1,0 +1,127 @@
+package sebak
+
+import (
+	"boscoin.io/sebak/lib/common"
+)
+
+type RoundVoteResult map[ /* Node.Address() */ string]sebakcommon.VotingHole
+
+type RoundVote struct {
+	SIGN   RoundVoteResult
+	ACCEPT RoundVoteResult
+}
+
+func NewRoundVote(ballot Ballot) (rv *RoundVote) {
+	rv = &RoundVote{
+		SIGN:   RoundVoteResult{},
+		ACCEPT: RoundVoteResult{},
+	}
+
+	rv.Vote(ballot)
+
+	return rv
+}
+
+func (rv *RoundVote) IsVoted(ballot Ballot) bool {
+	result := rv.GetResult(ballot.State())
+
+	_, found := result[ballot.Source()]
+	return found
+}
+
+func (rv *RoundVote) IsVotedByNode(state sebakcommon.BallotState, node string) bool {
+	result := rv.GetResult(state)
+
+	_, found := result[node]
+	return found
+}
+
+func (rv *RoundVote) Vote(ballot Ballot) (isNew bool, err error) {
+	if ballot.IsFromProposer() {
+		return
+	}
+
+	result := rv.GetResult(ballot.State())
+	_, isNew = result[ballot.Source()]
+	result[ballot.Source()] = ballot.Vote()
+
+	return
+}
+
+func (rv *RoundVote) GetResult(state sebakcommon.BallotState) (result RoundVoteResult) {
+	if !state.IsValidForVote() {
+		return
+	}
+
+	switch state {
+	case sebakcommon.BallotStateSIGN:
+		result = rv.SIGN
+	case sebakcommon.BallotStateACCEPT:
+		result = rv.ACCEPT
+	}
+
+	return result
+}
+
+func (rv *RoundVote) CanGetVotingResult(policy sebakcommon.VotingThresholdPolicy, state sebakcommon.BallotState) (RoundVoteResult, sebakcommon.VotingHole, bool) {
+	threshold := policy.Threshold(state)
+	if threshold < 1 {
+		return RoundVoteResult{}, sebakcommon.VotingNOTYET, false
+	}
+
+	result := rv.GetResult(state)
+	if len(result) < int(threshold) {
+		return result, sebakcommon.VotingNOTYET, false
+	}
+
+	var yes, no, expired int
+	for _, votingHole := range result {
+		switch votingHole {
+		case sebakcommon.VotingYES:
+			yes++
+		case sebakcommon.VotingNO:
+			no++
+		case sebakcommon.VotingEXP:
+			expired++
+		}
+	}
+
+	log.Debug(
+		"check threshold in isaac",
+		"threshold", threshold,
+		"yes", yes,
+		"no", no,
+		"expired", expired,
+		"policy", policy,
+		"state", state,
+	)
+
+	if state == sebakcommon.BallotStateSIGN {
+		if yes >= threshold {
+			return result, sebakcommon.VotingYES, true
+		} else if no >= threshold+1 {
+			return result, sebakcommon.VotingNO, true
+		}
+	} else if state == sebakcommon.BallotStateACCEPT {
+		if yes >= threshold {
+			return result, sebakcommon.VotingYES, true
+		} else if no >= threshold {
+			return result, sebakcommon.VotingNO, true
+		}
+	} else {
+		// do nothing
+	}
+
+	// check draw!
+	total := policy.Validators()
+	voted := yes + no + expired
+	if cannotBeOver(total-voted, threshold, yes, no) { // draw
+		return result, sebakcommon.VotingEXP, true
+	}
+
+	return result, sebakcommon.VotingNOTYET, false
+}
+
+func cannotBeOver(remain, threshold, yes, no int) bool {
+	return remain+yes < threshold && remain+no < threshold
+}
